Remove title from Modal and move it to Comparison subfolder
